+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Added VerificationHost.waitFor() utility method
   useful for convergence loops in tests.
 
+* Support ODATA $orderby and $top
+
 * Introduce ServiceOption.URI_NAMESPACE_OWNER. Allow
   for a service to handle all requests that start with
   a given prefix, allowing for emulation of other REST